forensic_hawkeye_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-forensic_hawkeye_env"
version = "0.1.0"
description = "Forensic accident reconstruction environment with 2.5D physics-based grading for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pymunk>=6.6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "forensic_hawkeye_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["forensic_hawkeye_env", "forensic_hawkeye_env.server", "forensic_hawkeye_env.server.scenarios"]
package-dir = { "forensic_hawkeye_env" = ".", "forensic_hawkeye_env.server" = "server", "forensic_hawkeye_env.server.scenarios" = "server/scenarios" }